+++ src/main/java/zaftnotameni/creatania/setup/FMLClientSetup.java
@@ -2,6 +2,7 @@ package zaftnotameni.creatania.setup;
 import net.minecraft.client.renderer.ItemBlockRenderTypes;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import zaftnotameni.creatania.ponder.CreataniaPonderIndex;
 import zaftnotameni.creatania.registry.Blocks;
 import zaftnotameni.creatania.util.Log;
 public class FMLClientSetup {
@@ -13,6 +14,8 @@ public class FMLClientSetup {
     ItemBlockRenderTypes.setRenderLayer(Blocks.CORRUPT_MANA_BLOCK.get(), RenderType.translucent());
     ItemBlockRenderTypes.setRenderLayer(Blocks.PURE_MANA_BLOCK.get(), RenderType.translucent());
     ItemBlockRenderTypes.setRenderLayer(Blocks.REAL_MANA_BLOCK.get(), RenderType.translucent());
+
+    CreataniaPonderIndex.register();
     Log.LOGGER.info("creatania client setup finished");
   }
 }